backend/requirements.txt
# ===========================================
# Nexus Backend - Python Dependencies
# ===========================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# LLM Providers
google-generativeai>=0.3.0
anthropic>=0.18.0
openai>=1.10.0

# Async & Utilities
httpx>=0.26.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# State Machine
transitions>=0.9.0

# WebSocket
websockets>=12.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
